chore(ui): remove setup-tracing from project creation
